[State Sync] Remove state sync v1!
