+++ aptos-node/Cargo.toml
@@ -53,8 +53,7 @@ inspection-service = { path = "../crates/inspection-service" }
 mempool-notifications = { path = "../state-sync/inter-component/mempool-notifications" }
 network = { path = "../network" }
 network-builder = { path = "../network/builder" }
-state-sync-multiplexer = { path = "../state-sync/state-sync-v2/state-sync-multiplexer" }
-state-sync-v1 = { path = "../state-sync/state-sync-v1" }
+state-sync-driver = { path = "../state-sync/state-sync-v2/state-sync-driver" }
 storage-interface = { path = "../storage/storage-interface" }
 storage-service-client = { path = "../state-sync/storage-service/client" }
 storage-service-server = { path = "../state-sync/storage-service/server" }
