+++ config/src/config/state_sync_config.rs
@@ -3,56 +3,13 @@
 
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
 #[serde(default, deny_unknown_fields)]
 pub struct StateSyncConfig {
-    // Size of chunk to request for state synchronization
-    pub chunk_limit: u64,
-    // The timeout of the state sync client to process a commit notification (in milliseconds)
-    pub client_commit_timeout_ms: u64,
-    // default timeout used for long polling to remote peer
-    pub long_poll_timeout_ms: u64,
-    // valid maximum chunk limit for sanity check
-    pub max_chunk_limit: u64,
-    // valid maximum timeout limit for sanity check
-    pub max_timeout_ms: u64,
-    // The timeout of the state sync coordinator to receive a commit ack from mempool (in milliseconds)
-    pub mempool_commit_timeout_ms: u64,
-    // default timeout to make state sync progress by sending chunk requests to a certain number of networks
-    // if no progress is made by sending chunk requests to a number of networks,
-    // the next sync request will be multicasted, i.e. sent to more networks
-    pub multicast_timeout_ms: u64,
-    // The timeout for ensuring sync requests are making progress (i.e., the maximum time between
-    // commits when processing a sync request).
-    pub sync_request_timeout_ms: u64,
-    // interval used for checking state synchronization progress
-    pub tick_interval_ms: u64,
-
-    // Everything above belongs to state sync v1 and will be removed in the future.
     pub data_streaming_service: DataStreamingServiceConfig,
     pub aptos_data_client: AptosDataClientConfig,
     pub state_sync_driver: StateSyncDriverConfig,
     pub storage_service: StorageServiceConfig,
-}
-
-impl Default for StateSyncConfig {
-    fn default() -> Self {
-        Self {
-            chunk_limit: 1000,
-            client_commit_timeout_ms: 5_000,
-            long_poll_timeout_ms: 10_000,
-            max_chunk_limit: 1000,
-            max_timeout_ms: 120_000,
-            mempool_commit_timeout_ms: 5_000,
-            multicast_timeout_ms: 30_000,
-            sync_request_timeout_ms: 60_000,
-            tick_interval_ms: 100,
-            data_streaming_service: DataStreamingServiceConfig::default(),
-            aptos_data_client: AptosDataClientConfig::default(),
-            state_sync_driver: StateSyncDriverConfig::default(),
-            storage_service: StorageServiceConfig::default(),
-        }
-    }
 }
 
 /// The bootstrapping mode determines how the node will bootstrap to the latest
@@ -100,7 +57,7 @@ impl ContinuousSyncingMode {
 #[serde(default, deny_unknown_fields)]
 pub struct StateSyncDriverConfig {
     pub bootstrapping_mode: BootstrappingMode, // The mode by which to bootstrap
-    pub enable_state_sync_v2: bool,            // If the node should sync with state sync v2
+    pub commit_notification_timeout_ms: u64, // The max time taken to process a commit notification
     pub continuous_syncing_mode: ContinuousSyncingMode, // The mode by which to sync after bootstrapping
     pub progress_check_interval_ms: u64, // The interval (ms) at which to check state sync progress
     pub max_connection_deadline_secs: u64, // The max time (secs) to wait for connections from peers
@@ -115,7 +72,7 @@ impl Default for StateSyncDriverConfig {
     fn default() -> Self {
         Self {
             bootstrapping_mode: BootstrappingMode::ApplyTransactionOutputsFromGenesis,
-            enable_state_sync_v2: true,
+            commit_notification_timeout_ms: 5000,
             continuous_syncing_mode: ContinuousSyncingMode::ApplyTransactionOutputs,
             progress_check_interval_ms: 100,
             max_connection_deadline_secs: 10,
