+++ crates/aptos-telemetry/Cargo.toml
@@ -34,4 +34,3 @@ aptosdb = { path = "../../storage/aptosdb" }
 consensus = { path = "../../consensus" }
 network = { path = "../../network" }
 state-sync-driver = { path = "../../state-sync/state-sync-v2/state-sync-driver" }
-state-sync-v1 = { path = "../../state-sync/state-sync-v1" }
