+++ crates/aptos-telemetry/src/core_metrics.rs
@@ -102,58 +102,39 @@ fn collect_state_sync_metrics(
     core_metrics: &mut BTreeMap<String, String>,
     node_config: &NodeConfig,
 ) {
-    // Depending on which state sync code is running, we need to
-    // grab the appropriate counters. Otherwise, the node will panic.
-    // TODO(joshlind): remove this when v1 is gone!
     let state_sync_driver_config = node_config.state_sync.state_sync_driver;
-    let is_state_sync_v2 = state_sync_driver_config.enable_state_sync_v2;
 
     // Get the state sync code version
-    let state_sync_code_version = if !is_state_sync_v2 { "1" } else { "2" };
-    core_metrics.insert(
-        STATE_SYNC_CODE_VERSION.into(),
-        state_sync_code_version.into(),
-    );
+    core_metrics.insert(STATE_SYNC_CODE_VERSION.into(), "2".into());
 
-    // Get the synced versions and syncing modes
-    if !is_state_sync_v2 {
-        core_metrics.insert(
-            STATE_SYNC_SYNCED_VERSION.into(),
-            state_sync_v1::counters::VERSION
-                .with_label_values(&["synced"])
-                .get()
-                .to_string(),
-        );
-    } else {
-        core_metrics.insert(
-            STATE_SYNC_SYNCED_EPOCH.into(),
-            state_sync_driver::metrics::STORAGE_SYNCHRONIZER_OPERATIONS
-                .with_label_values(&[StorageSynchronizerOperations::SyncedEpoch.get_label()])
-                .get()
-                .to_string(),
-        );
-        core_metrics.insert(
-            STATE_SYNC_SYNCED_VERSION.into(),
-            state_sync_driver::metrics::STORAGE_SYNCHRONIZER_OPERATIONS
-                .with_label_values(&[StorageSynchronizerOperations::Synced.get_label()])
-                .get()
-                .to_string(),
-        );
-        core_metrics.insert(
-            STATE_SYNC_BOOTSTRAP_MODE.into(),
-            state_sync_driver_config
-                .bootstrapping_mode
-                .to_label()
-                .into(),
-        );
-        core_metrics.insert(
-            STATE_SYNC_CONTINUOUS_SYNC_MODE.into(),
-            state_sync_driver_config
-                .continuous_syncing_mode
-                .to_label()
-                .into(),
-        );
-    }
+    core_metrics.insert(
+        STATE_SYNC_SYNCED_EPOCH.into(),
+        state_sync_driver::metrics::STORAGE_SYNCHRONIZER_OPERATIONS
+            .with_label_values(&[StorageSynchronizerOperations::SyncedEpoch.get_label()])
+            .get()
+            .to_string(),
+    );
+    core_metrics.insert(
+        STATE_SYNC_SYNCED_VERSION.into(),
+        state_sync_driver::metrics::STORAGE_SYNCHRONIZER_OPERATIONS
+            .with_label_values(&[StorageSynchronizerOperations::Synced.get_label()])
+            .get()
+            .to_string(),
+    );
+    core_metrics.insert(
+        STATE_SYNC_BOOTSTRAP_MODE.into(),
+        state_sync_driver_config
+            .bootstrapping_mode
+            .to_label()
+            .into(),
+    );
+    core_metrics.insert(
+        STATE_SYNC_CONTINUOUS_SYNC_MODE.into(),
+        state_sync_driver_config
+            .continuous_syncing_mode
+            .to_label()
+            .into(),
+    );
 }
 
 /// Collects the storage metrics and appends it to the given map
