--- state-sync/state-sync-v1/Cargo.toml
@@ -1,74 +1,0 @@
-[package]
-name = "state-sync-v1"
-version = "0.1.0"
-authors = ["Aptos Labs <opensource@aptoslabs.com>"]
-description = "Aptos state synchronizer v1 (state sync v1)"
-repository = "https://github.com/aptos-labs/aptos-core"
-homepage = "https://aptoslabs.com"
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-async-trait = "0.1.53"
-bcs = "0.1.3"
-channel = { path = "../../crates/channel" }
-fail = "0.5.0"
-futures = "0.3.21"
-itertools = { version = "0.10.0", default-features = false }
-once_cell = "1.10.0"
-proptest = { version = "1.0.0", optional = true }
-rand = "0.7.3"
-serde = { version = "1.0.137", default-features = false }
-thiserror = "1.0.31"
-tokio = { version = "1.18.2", features = ["full"] }
-tokio-stream = "0.1.8"
-
-aptos-config = { path = "../../config" }
-aptos-crypto = { path = "../../crates/aptos-crypto" }
-aptos-infallible = { path = "../../crates/aptos-infallible" }
-aptos-logger = { path = "../../crates/aptos-logger" }
-aptos-metrics-core = { path = "../../crates/aptos-metrics-core" }
-aptos-proptest-helpers = { path = "../../crates/aptos-proptest-helpers", optional = true }
-aptos-temppath = { path = "../../crates/aptos-temppath" }
-aptos-types = { path = "../../types" }
-aptos-vm = { path = "../../aptos-move/aptos-vm" }
-
-aptosdb = { path = "../../storage/aptosdb", optional = true }
-consensus-notifications = { path = "../inter-component/consensus-notifications" }
-event-notifications = { path = "../../state-sync/inter-component/event-notifications" }
-executor = { path = "../../execution/executor" }
-executor-test-helpers = { path = "../../execution/executor-test-helpers", optional = true }
-executor-types = { path = "../../execution/executor-types" }
-mempool-notifications = { path = "../inter-component/mempool-notifications" }
-memsocket = { path = "../../network/memsocket", optional = true }
-netcore = { path = "../../network/netcore" }
-network = { path = "../../network" }
-short-hex-str = { path = "../../crates/short-hex-str" }
-storage-interface = { path = "../../storage/storage-interface" }
-vm-genesis = { path = "../../aptos-move/vm-genesis", optional = true }
-
-[dev-dependencies]
-bytes = "1.1.0"
-claim = "0.5.0"
-proptest = "1.0.0"
-
-aptos-crypto = { path = "../../crates/aptos-crypto" }
-aptos-genesis = { path = "../../crates/aptos-genesis", features = ["testing"] }
-aptos-mempool = { path = "../../mempool", features = ["fuzzing"] }
-aptos-proptest-helpers = { path = "../../crates/aptos-proptest-helpers" }
-aptos-time-service = { path = "../../crates/aptos-time-service", features = ["testing"] }
-aptos-transaction-builder = { path = "../../sdk/transaction-builder" }
-aptosdb = { path = "../../storage/aptosdb" }
-channel = { path = "../../crates/channel" }
-executor-test-helpers = { path = "../../execution/executor-test-helpers" }
-memsocket = { path = "../../network/memsocket" }
-move-deps = { path = "../../aptos-move/move-deps", features = ["address32"] }
-network = { path = "../../network", features = ["fuzzing", "testing"] }
-network-builder = { path = "../../network/builder" }
-vm-genesis = { path = "../../aptos-move/vm-genesis", features = ["fuzzing"] }
-
-[features]
-default = []
-failpoints = ["fail/failpoints"]
-fuzzing = ["vm-genesis", "proptest", "executor-test-helpers", "aptos-config/fuzzing", "aptos-mempool/fuzzing", "aptos-types/fuzzing", "aptos-proptest-helpers", "memsocket/fuzzing"]
