+++ state-sync/state-sync-v2/state-sync-driver/src/driver_factory.rs
@@ -13,12 +13,14 @@ use crate::{
 use aptos_config::config::NodeConfig;
 use aptos_data_client::aptosnet::AptosNetDataClient;
 use aptos_infallible::Mutex;
+use aptos_types::move_resource::MoveStorage;
 use aptos_types::waypoint::Waypoint;
 use consensus_notifications::ConsensusNotificationListener;
 use data_streaming_service::streaming_client::StreamingServiceClient;
-use event_notifications::EventSubscriptionService;
+use event_notifications::{EventNotificationSender, EventSubscriptionService};
 use executor_types::ChunkExecutorTrait;
 use futures::channel::mpsc;
+use futures::executor::block_on;
 use mempool_notifications::MempoolNotificationSender;
 use std::sync::Arc;
 use storage_interface::DbReaderWriter;
@@ -43,10 +45,25 @@ impl DriverFactory {
         chunk_executor: Arc<ChunkExecutor>,
         mempool_notification_sender: MempoolNotifier,
         consensus_listener: ConsensusNotificationListener,
-        event_subscription_service: EventSubscriptionService,
+        mut event_subscription_service: EventSubscriptionService,
         aptos_data_client: AptosNetDataClient,
         streaming_service_client: StreamingServiceClient,
     ) -> Self {
+        // Notify subscribers of the initial on-chain config values
+        match (&*storage.reader).fetch_latest_state_checkpoint_version() {
+            Ok(synced_version) => {
+                if let Err(error) =
+                    event_subscription_service.notify_initial_configs(synced_version)
+                {
+                    panic!(
+                        "Failed to notify subscribers of initial on-chain configs: {:?}",
+                        error
+                    )
+                }
+            }
+            Err(error) => panic!("Failed to fetch the initial synced version: {:?}", error),
+        }
+
         // Create the notification handlers
         let (client_notification_sender, client_notification_receiver) = mpsc::unbounded();
         let client_notification_listener =
@@ -123,5 +140,37 @@ impl DriverFactory {
     /// Returns a new client that can be used to communicate with the driver
     pub fn create_driver_client(&self) -> DriverClient {
         DriverClient::new(self.client_notification_sender.clone())
+    }
+}
+
+/// A struct for holding the various runtimes required by state sync v2.
+/// Note: it's useful to maintain separate runtimes because the logger
+/// can prepend all logs with the runtime thread name.
+pub struct StateSyncRuntimes {
+    _aptos_data_client: Runtime,
+    state_sync: DriverFactory,
+    _storage_service: Runtime,
+    _streaming_service: Runtime,
+}
+
+impl StateSyncRuntimes {
+    pub fn new(
+        aptos_data_client: Runtime,
+        state_sync: DriverFactory,
+        storage_service: Runtime,
+        streaming_service: Runtime,
+    ) -> Self {
+        Self {
+            _aptos_data_client: aptos_data_client,
+            state_sync,
+            _storage_service: storage_service,
+            _streaming_service: streaming_service,
+        }
+    }
+
+    pub fn block_until_initialized(&self) {
+        let state_sync_client = self.state_sync.create_driver_client();
+        block_on(state_sync_client.notify_once_bootstrapped())
+            .expect("State sync v2 initialization failure");
     }
 }
