+++ state-sync/state-sync-v2/state-sync-driver/src/tests/driver.rs
@@ -14,7 +14,6 @@ use aptos_infallible::RwLock;
 use aptos_time_service::TimeService;
 use aptos_types::{
     event::EventKey,
-    move_resource::MoveStorage,
     on_chain_config::{new_epoch_event_key, ON_CHAIN_CONFIG_REGISTRY},
     transaction::{Transaction, WriteSetPayload},
     waypoint::Waypoint,
@@ -25,8 +24,7 @@ use claim::{assert_err, assert_none};
 use consensus_notifications::{ConsensusNotificationSender, ConsensusNotifier};
 use data_streaming_service::streaming_client::new_streaming_service_client_listener_pair;
 use event_notifications::{
-    EventNotificationListener, EventNotificationSender, EventSubscriptionService,
-    ReconfigNotificationListener,
+    EventNotificationListener, EventSubscriptionService, ReconfigNotificationListener,
 };
 use executor::chunk_executor::ChunkExecutor;
 use executor_test_helpers::bootstrap_genesis;
@@ -34,7 +32,7 @@ use futures::{FutureExt, StreamExt};
 use mempool_notifications::MempoolNotificationListener;
 use network::application::{interface::MultiNetworkSender, storage::PeerMetadataStorage};
 use std::{collections::HashMap, sync::Arc};
-use storage_interface::{DbReader, DbReaderWriter};
+use storage_interface::DbReaderWriter;
 use storage_service_client::StorageServiceClient;
 
 // TODO(joshlind): extend these tests to cover more functionality!
@@ -229,7 +227,7 @@ async fn create_driver_for_tests(
     // Create test aptos database
     let db_path = aptos_temppath::TempPath::new();
     db_path.create_as_dir().unwrap();
-    let (db, db_rw) = DbReaderWriter::wrap(AptosDB::new_for_test(db_path.path()));
+    let (_, db_rw) = DbReaderWriter::wrap(AptosDB::new_for_test(db_path.path()));
 
     // Bootstrap the genesis transaction
     let (genesis, _) = vm_genesis::test_genesis_change_set_and_validators(Some(1));
@@ -237,8 +235,6 @@ async fn create_driver_for_tests(
     bootstrap_genesis::<AptosVM>(&db_rw, &genesis_txn).unwrap();
 
     // Create the event subscription service and subscribe to events and reconfigurations
-    let storage: Arc<dyn DbReader> = db;
-    let synced_version = (&*storage).fetch_latest_state_checkpoint_version().unwrap();
     let mut event_subscription_service = EventSubscriptionService::new(
         ON_CHAIN_CONFIG_REGISTRY,
         Arc::new(RwLock::new(db_rw.clone())),
@@ -251,12 +247,6 @@ async fn create_driver_for_tests(
     let event_subscriber = event_subscription_service
         .subscribe_to_events(event_key_subscriptions)
         .unwrap();
-
-    // Notify subscribers of the initial configs
-    event_subscription_service
-        .notify_initial_configs(synced_version)
-        .unwrap();
-    reconfiguration_subscriber.select_next_some().await;
 
     // Create consensus and mempool notifiers and listeners
     let (consensus_notifier, consensus_listener) =
@@ -297,6 +287,10 @@ async fn create_driver_for_tests(
         aptos_data_client,
         streaming_service_client,
     );
+
+    // The driver will notify reconfiguration subscribers of the initial configs.
+    // Verify we've received this notification.
+    reconfiguration_subscriber.select_next_some().await;
 
     (
         driver_factory,
