+++ state-sync/state-sync-v2/state-sync-driver/src/tests/mod.rs
@@ -4,6 +4,7 @@
 mod bootstrapper;
 mod continuous_syncer;
 mod driver;
+mod driver_factory;
 mod mocks;
 mod storage_synchronizer;
 mod utils;
