--- state-sync/state-sync-v2/state-sync-multiplexer/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "state-sync-multiplexer"
-version = "0.1.0"
-authors = ["Aptos Labs <opensource@aptoslabs.com>"]
-repository = "https://github.com/aptos-labs/aptos-core"
-description = "A multiplexer supporting several different state sync versions"
-homepage = "https://aptoslabs.com"
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-futures = "0.3.21"
-tokio = { version = "1.18.2", features = ["full"] }
-
-aptos-config = { path = "../../../config" }
-aptos-data-client = { path = "../../aptos-data-client" }
-aptos-types = { path = "../../../types" }
-
-consensus-notifications = { path = "../../../state-sync/inter-component/consensus-notifications" }
-data-streaming-service = { path = "../data-streaming-service" }
-event-notifications = { path = "../../../state-sync/inter-component/event-notifications" }
-executor-types = { path = "../../../execution/executor-types" }
-mempool-notifications = { path = "../../../state-sync/inter-component/mempool-notifications" }
-network = { path = "../../../network" }
-state-sync-driver = { path = "../../../state-sync/state-sync-v2/state-sync-driver" }
-state-sync-v1 = { path = "../../../state-sync/state-sync-v1" }
-storage-interface = { path = "../../../storage/storage-interface" }
-
-[dev-dependencies]
-aptos-crypto = { path = "../../../crates/aptos-crypto" }
-aptos-genesis = { path = "../../../crates/aptos-genesis", features = ["testing"] }
-aptos-infallible = { path = "../../../crates/aptos-infallible" }
-aptos-temppath = { path = "../../../crates/aptos-temppath" }
-aptos-time-service = { path = "../../../crates/aptos-time-service", features = ["async", "testing"] }
-aptos-vm = { path = "../../../aptos-move/aptos-vm" }
-aptosdb = { path = "../../../storage/aptosdb" }
-executor = { path = "../../../execution/executor" }
-executor-test-helpers = { path = "../../../execution/executor-test-helpers" }
-mempool-notifications = { path = "../../../state-sync/inter-component/mempool-notifications" }
-storage-service-client = { path = "../../storage-service/client" }
