+++ terraform/helm/aptos-node/files/configs/fullnode.yaml
@@ -6,10 +6,6 @@ base:
 execution:
   genesis_file_location: /opt/aptos/genesis/genesis.blob
 
-state_sync:
-  state_sync_driver:
-    enable_state_sync_v2: {{ $.Values.validator.config.enable_state_sync_v2 }}
-
 storage:
   backup_service_address: "0.0.0.0:6186"
   storage_pruner_config:
