+++ terraform/helm/aptos-node/files/configs/validator.yaml
@@ -27,10 +27,6 @@ consensus:
   quorum_store_poll_count: {{ $.Values.validator.config.quorum_store_poll_count}}
   {{- end }}
 
-state_sync:
-  state_sync_driver:
-    enable_state_sync_v2: {{ $.Values.validator.config.enable_state_sync_v2 }}
-
 storage:
   storage_pruner_config:
     ledger_prune_window: {{ int $.Values.validator.config.ledger_prune_window }}
