+++ terraform/helm/aptos-node/values.yaml
@@ -35,7 +35,6 @@ validator:
     provide_genesis: true
     sync_only: false
     concurrency_level: 8
-    enable_state_sync_v2: true
     round_initial_timeout_ms:
     quorum_store_poll_count: 1
     ledger_prune_window: 10000000
