+++ testsuite/aptos-fuzzer/Cargo.toml
@@ -40,7 +40,6 @@ move-deps = { path = "../../aptos-move/move-deps", features = ["fuzzing"] }
 network = { path = "../../network", features = ["fuzzing"] }
 safety-rules = { path = "../../consensus/safety-rules", features = ["fuzzing", "testing"] }
 scratchpad = { path = "../../storage/scratchpad", features = ["fuzzing"] }
-state-sync-v1 = { path = "../../state-sync/state-sync-v1", features = ["fuzzing", "aptosdb"] }
 storage-interface = { path = "../../storage/storage-interface" }
 
 [dev-dependencies]
